Refactor file handling logic to utility file
Extracted `fileListFromFileArray`, `convertFileHandlesToFileList`, and `download` functions to `src/lib/file-utils.ts` to reduce duplication and improve maintainability.
Updated `src/routes/+page.svelte` and `src/routes/decrypt/+page.svelte` to use the shared utility.
